Moving fps and cur to scikit-cosmo version with backwards compatibility
