+++ tests/python/test_filter.py
@@ -1,0 +1,10 @@
+import unittest
+
+from bindings.rascal.utils import FPSFilter, CURFilter
+
+class CURTest(unittest.TestCase):
+    def test_one(self):
+        pass
+
+if __name__ == "__main__":
+    unittest.main(verbosity=2)
